+++ jwst/cube_build/cube_internal_cal.py
@@ -3,8 +3,8 @@ the interpolation method = area , coord_system = internal_cal
 """
 import numpy as np
 from stdatamodels.jwst.datamodels import dqflags
+from stdatamodels.jwst.transforms.models import _toindex
 
-from jwst.transforms.models import _toindex
 from .cube_match_internal import cube_wrapper_internal  # c extension